Migrate pages/api/ over to TS
