+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@fullhuman/postcss-purgecss": "^2.2.0",
     "@types/cheerio": "^0.22.18",
+    "@types/cors": "^2.8.6",
     "@types/mongodb": "^3.5.20",
     "autoprefixer": "^9.8.0",
     "cheerio": "^1.0.0-rc.3",
